+++ examples/vanilla/vite.config.ts
@@ -4,7 +4,7 @@ import { resolve } from "path"
 export default defineConfig({
   resolve: {
     alias: {
-      "@mymo/avatar": resolve(__dirname, "../../packages/core/src/index.ts"),
+      "@mymosdk/avatar": resolve(__dirname, "../../packages/core/src/index.ts"),
     },
   },
 })